telegram-bot/requirements.txt
python-telegram-bot==21.7
aiohttp>=3.9.0
python-dotenv>=1.0.0
gspread>=6.1.2
google-auth>=2.35.0
